+++ src/pooja/forloop/Assignment_14.java
@@ -6,7 +6,7 @@ Input : Range -> 1 to 60
 Output : In the range 1 to 70, calculated sum is 150
          In the range 1 to 70, 4 numbers are divisible by 3 and 5.
 */
-package pooja;
+package pooja.forloop;
 
 public class Assignment_14 {
